prop interface definition in the files they belong to, no imports
